+++ app/src/main/java/pro/haichuang/tzs144/model/OrderDetailModel.java
@@ -277,16 +277,16 @@ public class OrderDetailModel {
             private int orderGoodsId;
             private int waterGoodsId;
             private String waterName;
-            private float waterNum;
-            private float waterDeductNum;
+            private int waterNum;
+            private int waterDeductNum;
             private String goodsSpecsName;
             private float goodsPrice;
-            private float couponDeductNum;
+            private int couponDeductNum;
             private String couponImg;
             private String goodsName;
             private float goodsNum;
             private String monthImg;
-            private float monthDeductNum;
+            private int monthDeductNum;
             private List<String> materialList;
             private List<BindMaterList> bindMaterList;
             private boolean showWater;
@@ -361,19 +361,19 @@ public class OrderDetailModel {
                 this.bindMaterList = bindMaterList;
             }
 
-            public float getWaterNum() {
+            public int getWaterNum() {
                 return waterNum;
             }
 
-            public void setWaterNum(float waterNum) {
+            public void setWaterNum(int waterNum) {
                 this.waterNum = waterNum;
             }
 
-            public float getWaterDeductNum() {
+            public int getWaterDeductNum() {
                 return waterDeductNum;
             }
 
-            public void setWaterDeductNum(float waterDeductNum) {
+            public void setWaterDeductNum(int waterDeductNum) {
                 this.waterDeductNum = waterDeductNum;
             }
 
@@ -385,11 +385,11 @@ public class OrderDetailModel {
                 this.goodsPrice = goodsPrice;
             }
 
-            public float getCouponDeductNum() {
+            public int getCouponDeductNum() {
                 return couponDeductNum;
             }
 
-            public void setCouponDeductNum(float couponDeductNum) {
+            public void setCouponDeductNum(int couponDeductNum) {
                 this.couponDeductNum = couponDeductNum;
             }
 
@@ -425,11 +425,11 @@ public class OrderDetailModel {
                 this.monthImg = monthImg;
             }
 
-            public float getMonthDeductNum() {
+            public int getMonthDeductNum() {
                 return monthDeductNum;
             }
 
-            public void setMonthDeductNum(float monthDeductNum) {
+            public void setMonthDeductNum(int monthDeductNum) {
                 this.monthDeductNum = monthDeductNum;
             }
